test branch for kuberenetes deployment

+++ scripts/s3-support-bundles/s3_bundle_generate.sh
@@ -82,9 +82,15 @@ s3_bundle_location=$bundle_path/s3
 
 haproxy_config="/etc/haproxy/haproxy.cfg"
 # Collecting rotated logs for haproxy and ldap along with live log
+haproxy_status_log="/var/log/cortx/haproxy-status.log"
+haproxy_log="/var/log/cortx/haproxy.log"
+ldap_log="/var/log/cortx/slapd.log"
+
 haproxy_log="$base_log_file_path/haproxy.log"
 haproxy_status_log="$base_log_file_path/haproxy-status.log"
-haproxy_log_k8s=$(s3confstore "yaml:///opt/seagate/cortx/s3/mini-prov/s3_prov_config.yaml" getkey --key="S3_HAPROXY_LOG_SYMLINK")
+
+haproxy_sysconfig=$(s3confstore "yaml:///opt/seagate/cortx/s3/mini-prov/s3_prov_config.yaml" getkey --key="S3_HAPROXY_SYSCONF_SYMLINK")
+haproxy_sysconfig_log_file=$(s3confstore "properties://$haproxy_sysconfig" getkey --key="LOG_FILE")
 
 s3server_config="$base_config_file_path/s3/conf/s3config.yaml"
 authserver_config="$base_config_file_path/auth/resources/authserver.properties"
@@ -118,6 +124,7 @@ first_s3_m0trace_file="$tmp_dir/first_s3_m0trace_file"
 m0trace_files_count=5
 s3_core_files_max_count=11
 max_allowed_core_size=10737418240 #e.g 10GB byte value
+compress_core_file=true
 
 # LDAP data
 ldap_dir="$tmp_dir/ldap"
@@ -151,13 +158,15 @@ s3_motr_dir=$(s3confstore "yaml://$s3server_config" getkey --key="S3_SERVER_CONF
 # from s3_core_dir directory, if available
 collect_core_files(){
   echo "Collecting core files..."
-  core_filename_pattern="core-s3server.*.gz"
+  # core_filename_pattern="core-s3server.*.gz"
+  core_filename_pattern="core-s3server.*"
   mkdir -p $s3_core_files
   cwd=$(pwd)
   if [ ! -d "$s3_core_dir" ];
   then
       return;
   fi
+
   cd $s3_core_dir
   # get recent modified core files from directory
   (ls -t $core_filename_pattern 2>/dev/null | head -$s3_core_files_max_count) | xargs -I '{}' cp '{}' $s3_core_files
@@ -170,47 +179,56 @@ collect_core_files(){
       # iterate over the s3_core_files directory and extract call stack for each file
       for s3corefile in "$s3_core_files"/*
       do
-         zipped_core_name=$(basename "$s3corefile")
-         core_name=${zipped_core_name%".gz"}
+         #zipped_core_name=$(basename "$s3corefile")
+         #core_name=${zipped_core_name%".gz"}
+         core_name=$(basename "$s3corefile")
          callstack_file="$s3_core_files"/callstack."$core_name".txt
          # Check the uncompressed size of gz core file
-         core_size=$(gunzip -lq $s3corefile)
-         uncompressed_size=$(cut -d' ' -f2 <<< $core_size)
+         #core_size=$(gunzip -lq $s3corefile)
+         #uncompressed_size=$(cut -d' ' -f2 <<< $core_size)
+
+         uncompressed_size=$(du -k "$s3corefile" | cut -f1)
          if [ $uncompressed_size -gt $max_allowed_core_size ];
          then
              echo "Ignoring $core_name because of core size($uncompressed_size Bytes) is greater than 10GB(10737418240 Bytes)" >> ./ignored_core_files.txt
              rm -f "$s3corefile"
          else
-             # unzip the core file, to read it using gdb
-             gunzip $s3corefile 2>/dev/null
              printf "Core name: $core_name\n" >> "$callstack_file"
              printf "Callstack:\n\n" >> "$callstack_file"
-             # generate gdb bt and append into the callstack_file
-             gdb --batch --quiet -ex "thread apply all bt full" -ex "quit" $s3server_binary\
-             "$s3_core_files/$core_name" 2>/dev/null >> "$callstack_file"
-             printf "\n**************************************************************************\n" >> "$callstack_file"
+             # compress and collect core file
+             if [ $compress_core_file == true ];
+             then
+                 echo "Compress and collect s3 core file: $core_name"
+                 gzip --fast $s3corefile 2>/dev/null
+             else
+                 gunzip $s3corefile 2>/dev/null
+                 # unzip the core file, to read it using gdb
+                 # generate gdb bt and append into the callstack_file
+                 gdb --batch --quiet -ex "thread apply all bt full" -ex "quit" $s3server_binary\
+                 "$s3_core_files/$core_name" 2>/dev/null >> "$callstack_file"
+                 printf "\n**************************************************************************\n" >> "$callstack_file"
 
-             # generate gdb registers dump
-             printf "Register state:\n\n" >> "$callstack_file"
-             gdb --batch --quiet -ex "info registers" -ex "quit" $s3server_binary\
-             "$s3_core_files/$core_name" 2>/dev/null >> "$callstack_file"
-             printf "\n**************************************************************************\n" >> "$callstack_file"
+                 # generate gdb registers dump
+                 printf "Register state:\n\n" >> "$callstack_file"
+                 gdb --batch --quiet -ex "info registers" -ex "quit" $s3server_binary\
+                 "$s3_core_files/$core_name" 2>/dev/null >> "$callstack_file"
+                 printf "\n**************************************************************************\n" >> "$callstack_file"
+                 # generate gdb assembly code
+                 printf "Assembly code:\n\n" >> "$callstack_file"
+                 gdb --batch --quiet -ex "disassemble" -ex "quit" $s3server_binary\
+                 "$s3_core_files/$core_name" 2>/dev/null >> "$callstack_file"
+                 printf "\n**************************************************************************\n" >> "$callstack_file"
 
-             # generate gdb assembly code
-             printf "Assembly code:\n\n" >> "$callstack_file"
-             gdb --batch --quiet -ex "disassemble" -ex "quit" $s3server_binary\
-             "$s3_core_files/$core_name" 2>/dev/null >> "$callstack_file"
-             printf "\n**************************************************************************\n" >> "$callstack_file"
+                 # generate gdb memory map
+                 printf "Memory map:\n\n" >> "$callstack_file"
+                 gdb --batch --quiet -ex "info proc mapping" -ex "quit" $s3server_binary\
+                 "$s3_core_files/$core_name" 2>/dev/null >> "$callstack_file"
+                 printf "\n**************************************************************************\n" >> "$callstack_file"
 
-             # generate gdb memory map
-             printf "Memory map:\n\n" >> "$callstack_file"
-             gdb --batch --quiet -ex "info proc mapping" -ex "quit" $s3server_binary\
-             "$s3_core_files/$core_name" 2>/dev/null >> "$callstack_file"
-             printf "\n**************************************************************************\n" >> "$callstack_file"
-
-             # delete the inflated core file
-             rm -f "$s3_core_files/$core_name"
-             # delete the zipped core file
+                 # delete the inflated core file
+                 rm -f "$s3_core_files/$core_name"
+             fi
+             # delete the copied core file
              rm -f "$s3corefile"
          fi
       done
@@ -395,19 +413,6 @@ then
     args+=($s3cluster_config)
 fi
 
-# Collect s3cluster config file if available
-if [ -f "$s3cluster_config" ];
-then
-    args=$args" "$s3cluster_config
-fi
-
-# Collect s3cluster config file if available
-if [ -f "$s3cluster_config" ];
-then
-    args=$args" "$s3cluster_config
-fi
-
-# Collect s3startsystem script file if available
 if [ -f "$s3startsystem_script" ];
 then
     args+=($s3startsystem_script)
@@ -562,4 +567,3 @@ tar -cvJf $s3_bundle_location/$bundle_name "${args[@]}" --warning=no-file-change
 # Clean up temp files
 cleanup_tmp_files
 echo "S3 support bundle generated successfully at $s3_bundle_location/$bundle_name !!!"
-
